pyproject.toml
[tool.poetry]
name = "taipei-day-trip"
version = "0.1.0"
description = ""
authors = ["lucia <thisislucia1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
mysql-connector-python = "^8.4.0"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
requests = "^2.32.3"
pillow = "^10.3.0"
cloudinary = "^1.40.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
